requirements.txt
Django~=5.2
environs~=14.1.1
psycopg~=3.2.6
python-slugify~=8.0.4